Fix test at `test_geometry_serializer_method_field_nogeo`
The function `add_to_fields` does not have any checks for `None` and
since the app now supports `geo_field = None` it adds it to the fields
of the serializer. In any case, adding an `assert` in the `add_to_fields`
function would have made the error obvious in the first place.